+++ apostello/management/commands/setup_periodic_tasks.py
@@ -26,6 +26,14 @@ class Command(BaseCommand):
         next_3am = add_day_if_req(now.replace(hour=3))
         next_230am = add_day_if_req(now.replace(hour=2, minute=30))
         next_2130 = add_day_if_req(now.replace(hour=21, minute=30))
+        if Schedule.objects.filter(func='apostello.tasks.send_queued_sms'
+                                   ).count() < 1:
+            Schedule.objects.create(
+                func='apostello.tasks.send_queued_sms',
+                schedule_type=Schedule.MINUTES,
+                minutes=1,
+            )
+
         if Schedule.objects.filter(func='apostello.tasks.pull_elvanto_groups'
                                    ).count() < 1:
             Schedule.objects.create(