+++ assets/js/components/group_row.jsx
@@ -5,7 +5,7 @@ const GroupRow = props => (
   <tr>
     <td><a href={props.group.url}>{props.group.name}</a></td>
     <td>{props.group.description}</td>
-    <td className="collapsing">{`\$${props.group.cost}`}</td>
+    <td className="collapsing">{`$${props.group.cost}`}</td>
     <td className="collapsing">
       <ArchiveButton item={props.group} archiveFn={props.archiveGroup} />
     </td>